requirements.txt
python-dotenv
markdown
pydantic
requests
python-frontmatter